Build configuration file

Compile metadata from all MRR and interogation sites from PTAGIS

buildConfig(
  node_assign = c("array", "site", "antenna"),
  array_suffix = c("UD", "UMD", "A0B0")
)

Source

http://www.ptagis.org

Arguments

node_assign

should nodes by assigned by array, site or antenna? Default is array.

array_suffix

if node_assign = "array", should nodes uses the suffixes of _U & _D (UD), _U, _M, & _D (UMD), or _A0 & _B0 (A0B0) to denote upstream, middle, and downstream arrays, respectively. Default is UD. If array_suffix = "UD" or array_suffix = "A0B0", the middle array is grouped with the upstream array.

Author

Kevin See

Examples

buildConfig()
#> Querying INT sites' metadata
#> Querying INT sites' configuration information
#> Querying MRR sites' metadata
#> # A tibble: 7,571 × 15
#>    site_code config_id antenna_id node  start_date          end_date           
#>    <chr>         <dbl> <chr>      <chr> <dttm>              <dttm>             
#>  1 0HR             100 01         0HR   2022-07-26 00:00:00 NA                 
#>  2 0HR             100 02         0HR   2022-07-26 00:00:00 NA                 
#>  3 0HR             100 03         0HR   2022-07-26 00:00:00 NA                 
#>  4 158             100 D1         158_U 2011-11-29 00:00:00 NA                 
#>  5 158             100 D2         158_U 2011-11-29 00:00:00 NA                 
#>  6 158             100 D3         158_D 2011-11-29 00:00:00 NA                 
#>  7 158             100 D4         158_U 2011-11-29 00:00:00 NA                 
#>  8 158             100 D5         158_U 2011-11-29 00:00:00 NA                 
#>  9 158             100 D6         158_D 2011-11-29 00:00:00 NA                 
#> 10 15D             100 C1         15D_U 2012-03-15 00:00:00 2019-06-30 12:00:00
#> # ℹ 7,561 more rows
#> # ℹ 9 more variables: site_type <chr>, site_name <chr>, antenna_group <chr>,
#> #   site_description <chr>, site_type_name <chr>, rkm <chr>, rkm_total <dbl>,
#> #   latitude <dbl>, longitude <dbl>