Build Node Order

Creates a data frame showing the path (through various detection sites) to reach each detection location in a parent-child table, and the node order of that location

buildNodeOrder(parent_child = NULL, direction = "u")

Arguments

parent_child

data frame with at least `parent` and `child` columns. Can be created with `buildParentChild()` function.

direction

the direction of movement. Parent-child tables are generally built imagining parents as downstream from children. Assuming that's the case, should the paths be upstream (`u` or `upstream`), the default, or downstream (`d` or `downstream`)?

Value

data frame containing a column,`end_loc`, showing the possible final locations, a column, `node_order`, showing the order of that node along the path, and a column, `path`, showing all the detection locations to pass on the way to that final location.

Author

Kevin See

Examples

buildNodeOrder()
#> Error in buildPaths(parent_child = parent_child, direction = direction): !is.null(parent_child) is not TRUE