Determine next downstream site

Determine which detection site is the closest downstream site from a starting site, based on hydrographic flowlines

findDwnstrmSite(
  site_id = NULL,
  flow_lines = NULL,
  sites_joined = NULL,
  quiet = TRUE
)

Arguments

site_id

Site code for the starting detection site.

flow_lines

An `sf` object containing the flowlines (output of `queryFlowlines()`). Must have a column name called Hydroseq with unique identifiers, and a column name called DnHydroseq containing the hydro sequence of the next downstream segment

sites_joined

An `sf` class object containing points of all detection sites. Must contain a column named `site_code` containing the site code of each site, and a column called hydroseq containing the hydro sequence closest to each site.

Value

character of the site code downstream

Author

Kevin See