Process capture histories for Prosser Dam

Using a site/node configuration file, a list of modeled sites, a parent-child table of nodes, and the raw capture histories from PTAGIS, process them. This function is specific to fish tagged at Prosser dam.

processCapHist_PRO(
  start_date = NULL,
  tagging_site = c("PRO", "PROTRP", "PROSRD"),
  configuration = NULL,
  parent_child = NULL,
  observations = NULL,
  last_obs_date = NULL,
  truncate = T,
  site_df,
  step_num = 2,
  save_file = F,
  file_name = NULL
)

Arguments

start_date

observations that occurred before this date (YYYYMMDD format) will not be included.

tagging_site

vector of site codes used to determine when fish were caught in the fish trap

last_obs_date

Character string in the format "YYYYMMDD". If included, the output will suggest that observations after this date should be deleted.

save_file

Should output be written to an file? Default value is FALSE.

file_name

If save_file == TRUE, this is the file name (with possible extension) to be saved to.

Value

list

Author

Kevin See

Examples

processCapHist_PRO()
#> Error in processCapHist_PRO(): could not find function "processCapHist_PRO"