Process tags observed at Lower Granite and any subsequent detections upstream.

Using the raw import from DART's observation query, and the configuration files the function process the data into a PITcleanr capture history dataframe. A list is returned that contains the raw DART query export and the cleaned capture history dataset.

processDART_LGR(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  spawnYear = NULL,
  configuration = NULL,
  truncate = T
)

Arguments

species
Chinook, Coho, Steelhead and Sockeye

spawnYear
available years include spawn year 2010 to present

configuration
output from buildConfig function

truncate
should repeat observations at the same IPTDS node be removed from cleaned capture history dataset?

Author

Ryan N. Kinzer

Examples

my_config <- buildConfig()

#> Querying INT sites' metadata
#> Querying INT sites' configuration information
#> Querying MRR sites' metadata

dart_obs <- processDART_LGR(species = 'Chinook',
                           spawnYear = 2020,
                           configuration = my_config,
                           truncate = T)

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_character(),
#>   start_time = col_datetime(format = ""),
#>   obs_time = col_datetime(format = ""),
#>   shuc = col_double(),
#>   mark_date = col_date(format = ""),
#>   rel_date = col_date(format = ""),
#>   t_species = col_double(),
#>   t_run = col_double(),
#>   length = col_double(),
#>   rhuc = col_double(),
#>   trans_year = col_double(),
#>   config = col_double()
#> )
#> ℹ Use `spec()` for the full column specifications.

#> Error: Problem with `mutate()` input `ConfigID`.
#> x object 'ConfigID' not found
#> ℹ Input `ConfigID` is `as.character(ConfigID)`.