Quality Check Complete Tag Histories

The function takes complete capture histories downloaded from PTAGIS and performs several QA/QC checks on them. It returns the results in the form of a list.

qcTagHistory(ptagis_file = NULL, ignore_event_vs_release = F)

Arguments

ptagis_file
is the path to the PTAGIS observation file downloaded as a csv from PTAGIS. This must be the output from a Complete Tag History query (part of the Advanced Reporting). This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type, Event Site Code, Event Date Time, Antenna, Antenna Group Configuration, Event Release Site Code, and Event Release Date Time.

ignore_event_vs_release
Should the function attempt to choose whether to use the event time value or the event release time value for different release batches? Default is `FALSE`. If set to `TRUE`, the event time value will be used

Value

a list consisting of a vector of tag codes that are listed as "Disown" in PTAGIS (`disown_tags`), a vector of tag codes that are listed as "Orphan" in PTAGIS (`orphan_tags`), and, if `ignore_event_vs_release` is `TRUE`, a tibble containing information about sites with release information to help the user determine with event time or release time should be used for those sites.

Author

Kevin See

Examples

ptagis_file = system.file("extdata", "PRO_Steelhead_2019.csv", package = "PITcleanr")
qcTagHistory(ptagis_file)

#> $disown_tags
#> character(0)
#> 
#> $orphan_tags
#> character(0)
#> 
#> $rel_time_batches
#> # A tibble: 8 x 11
#>   mark_species_name  year event_site_type_des… event_type_name event_site_code_…
#>   <chr>             <dbl> <chr>                <chr>           <chr>            
#> 1 Coho               2017 Hatchery             Mark            EAGH             
#> 2 Coho               2018 Dam                  Recapture       PRO              
#> 3 Steelhead          2015 River                Mark            TOPPEC           
#> 4 Steelhead          2018 Dam                  Mark            PRO              
#> 5 Steelhead          2018 Dam                  Recapture       PRO              
#> 6 Steelhead          2019 Dam                  Mark            PRO              
#> 7 Steelhead          2019 Dam                  Recapture       PRO              
#> 8 Steelhead          2019 Dam                  Recapture       ROZ              
#> # … with 6 more variables: n_tags <int>, n_events <int>, n_release <int>,
#> #   rel_greq_event <int>, rel_ls_event <int>, event_rel_ratio <dbl>
#>