PTAGIS MRR Data File

Queries a raw MRR data file for PTAGIS MRR sites, based upon a known file name. NOTE: only the latest, corrected, version of a MRR file is returned

queryMRRDataFile(file_nm = NULL, file_path = NULL, text_only = FALSE)

Source

http://www.ptagis.org

Arguments

file_nm

PTAGIS file name.

file_path

if the user has downloaded the PTAGIS file, this function can read in the file, if the file path is included here. The default value, NULL, indicates that the file should be queried from the PTAGIS API interface.

text_only

should this function return a file only containing the text of the MRR file? Default is FALSE, which will parse the MRR data file into a data.frame.

Author

Kevin See

Examples

queryMRRDataFile(file_nm = "NBD-2014-188-PRD.xml")
#> # A tibble: 3,426 × 20
#>    capture_method conditional_comments event_date          event_site event_type
#>    <chr>          <chr>                <dttm>              <chr>      <chr>     
#>  1 LADDER         CW FE MT RE RF       2014-07-07 08:45:00 PRDLD1     Recapture 
#>  2 LADDER         RF MT AT FE          2014-07-07 07:45:00 PRDLD1     Mark      
#>  3 LADDER         RF MT AT MA          2014-07-07 07:45:00 PRDLD1     Mark      
#>  4 LADDER         RF MT AT FE          2014-07-07 07:45:00 PRDLD1     Mark      
#>  5 LADDER         RF MT AT FE          2014-07-07 07:45:00 PRDLD1     Mark      
#>  6 LADDER         RE RF MT MA          2014-07-07 07:45:00 PRDLD1     Recapture 
#>  7 LADDER         RF MT AT FE          2014-07-07 07:45:00 PRDLD1     Mark      
#>  8 LADDER         AD AT FE MT RF       2014-07-07 08:45:00 PRDLD1     Mark      
#>  9 LADDER         RF MT AT FE          2014-07-07 07:45:00 PRDLD1     Mark      
#> 10 LADDER         RF MT AT MA          2014-07-07 07:45:00 PRDLD1     Mark      
#> # ℹ 3,416 more rows
#> # ℹ 15 more variables: length <dbl>, life_stage <chr>, mark_temperature <chr>,
#> #   migration_year <chr>, organization <chr>, pit_tag <chr>,
#> #   release_date <dttm>, release_site <chr>, release_temperature <chr>,
#> #   sequence_number <dbl>, species_run_rear_type <chr>, tagger <chr>,
#> #   text_comments <chr>, mark_method <chr>, second_pit_tag <chr>