Read In PTAGIS Complete Tag History

The function reads in complete tag histories from PTAGIS that have been downloaded and coverts the date time values from character vectors into date-time vectors

readCTH(ptagis_file = NULL)

Arguments

ptagis_file
is the path to the PTAGIS observation file downloaded as a csv from PTAGIS. This must be the output from a Complete Tag History query (part of the Advanced Reporting). This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type, Event Site Code, Event Date Time, Antenna, Antenna Group Configuration, Event Release Site Code, and Event Release Date Time.

Value

a tibble containing the data downloaded from PTAGIS through a complete capture history query.

Author

Kevin See

Examples

ptagis_file = system.file("extdata", "PRO_Steelhead_2019.csv", package = "PITcleanr")
readCTH(ptagis_file)

#> # A tibble: 1,119 x 12
#>    tag_code  mark_species_name mark_rear_type_… event_type_name event_site_type…
#>    <chr>     <chr>             <chr>            <chr>           <chr>           
#>  1 384.3B23… Steelhead         Wild Fish or Na… Mark            River           
#>  2 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  3 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  4 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  5 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  6 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  7 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  8 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#>  9 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#> 10 384.3B23… Steelhead         Wild Fish or Na… Observation     Instream Remote…
#> # … with 1,109 more rows, and 7 more variables: event_site_code_value <chr>,
#> #   event_date_time_value <dttm>, antenna_id <chr>,
#> #   antenna_group_configuration_value <dbl>,
#> #   event_release_site_code_code <chr>, event_release_date_time_value <dttm>,
#> #   cth_count <dbl>