Compress Detections from DART Query

For certain locations (currently Lower Granite Dam and Priest Rapids Dam), query DART to return all PIT tags observed there from a given species in a given spawn year, as well as any subsequent detections upstream, and then compress them.

compressDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawn_year = NULL,
  configuration = NULL,
  ...
)

Arguments

species

Chinook, Coho, Steelhead and Sockeye

loc

code for initial location. Currently includes options for Lower Granite (GRA) and Priest Rapids (PRA) dams.

spawn_year

available years includes spawn year 2010 to present

configuration

is a data frame which assigns node names to unique site, antenna, and site configuration ID combinations. One example can be built with the function `buildConfig`. If no configuration file is provided, nodes are considered site codes by default. If nodes are assigned, the column name should be `node`.

...

other arguements to be passed to the `compress` function if the default values are not suitable.

Value

A tibble with particular information from a complete tag history PTAGIS query.

Author

Kevin See

Examples

compressDART(species = 'Chinook',
                       loc = "GRA",
                       spawn_year = 2019)
#> $compress_obs
#> # A tibble: 7,617 x 9
#>    tag_code       node    slot event_type_name n_dets min_det            
#>    <chr>          <chr>  <int> <chr>            <int> <dttm>             
#>  1 384.0A0301AFC3 GRA        1 Observation         11 2019-09-27 05:50:49
#>  2 384.1B79719E60 GRA        1 Observation         13 2019-06-08 11:45:02
#>  3 384.1B79719E60 ESS        2 Observation          2 2019-07-06 14:51:51
#>  4 384.1B79719E60 JOHNSC     3 Recapture           48 2019-07-08 14:36:02
#>  5 384.1B79775BEB GRA        1 Observation         17 2019-05-25 06:11:04
#>  6 384.3B23A4A244 GRA        1 Observation         15 2019-09-10 07:54:21
#>  7 384.3B23A562DE GRA        1 Observation          6 2019-06-07 11:23:51
#>  8 384.3B23A562DE LGRLDR     2 Recapture            1 2019-06-07 12:39:19
#>  9 384.3B23A562DE GRA        3 Observation         10 2019-06-07 14:30:16
#> 10 384.3B23A562DE ZEN        4 Observation          1 2019-06-30 00:36:40
#> # i 7,607 more rows
#> # i 3 more variables: max_det <dttm>, duration <drtn>, travel_time <drtn>
#> 
#> $dart_obs
#> # A tibble: 67,203 x 34
#>    tag_code       mark_species_name mark_rear_type_name event_type_name
#>    <chr>          <chr>             <chr>               <chr>          
#>  1 3DD.007784D29F Chinook           H                   Observation    
#>  2 3DD.007784D29F Chinook           H                   Observation    
#>  3 3DD.007784D29F Chinook           H                   Observation    
#>  4 3DD.007784D29F Chinook           H                   Observation    
#>  5 3DD.007784D29F Chinook           H                   Observation    
#>  6 3DD.007784D29F Chinook           H                   Recapture      
#>  7 3DD.007784D29F Chinook           H                   Observation    
#>  8 3DD.007784D29F Chinook           H                   Observation    
#>  9 3DD.007784D29F Chinook           H                   Observation    
#> 10 3DD.007784D29F Chinook           H                   Observation    
#> # i 67,193 more rows
#> # i 30 more variables: event_site_code_value <chr>,
#> #   event_date_time_value <dttm>, antenna_id <chr>,
#> #   antenna_group_configuration_value <dbl>, start_time <dttm>, stage <chr>,
#> #   shuc <dbl>, srkm <chr>, mark_date <date>, file_id <chr>, mark_site <chr>,
#> #   rel_site <chr>, rel_date <date>, t_species <dbl>, t_run <dbl>,
#> #   length <dbl>, esu_type <chr>, rhuc <dbl>, rrkm <chr>, flags <chr>, ...
#>