Query DART for observations of PIT tags

For certain locations (currently Lower Granite Dam and Priest Rapids Dam), query DART to return all PIT tags observed there from a given species in a given spawn year, as well as any subsequent detections upstream.

queryObsDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawn_year = NULL
)

Arguments

species

Chinook, Coho, Steelhead and Sockeye

loc

code for initial location. Currently includes options for Lower Granite (GRA) and Priest Rapids (PRA) dams.

spawn_year

available years includes spawn year 2010 to present

Value

A tibble with particular information from a complete tag history PTAGIS query.

Author

Kevin See

Examples

queryObsDART(species = 'Chinook',
                       loc = "GRA",
                       spawn_year = 2019)
#> # A tibble: 67,203 × 33
#>    tag_id      start_time          obs_time            obs_site coil_id obs_type
#>    <chr>       <dttm>              <dttm>              <chr>    <chr>   <chr>   
#>  1 3DD.007784… 2019-05-02 17:34:35 2019-05-02 17:34:35 GRA      B4      INT     
#>  2 3DD.007784… 2019-05-02 17:34:35 2019-05-02 18:25:00 GRA      16      INT     
#>  3 3DD.007784… 2019-05-02 17:34:35 2019-05-02 18:25:00 GRA      18      INT     
#>  4 3DD.007784… 2019-05-02 17:34:35 2019-05-02 18:25:00 GRA      12      INT     
#>  5 3DD.007784… 2019-05-02 17:34:35 2019-05-02 18:25:00 GRA      14      INT     
#>  6 3DD.007784… 2019-05-02 17:34:35 2019-05-03 08:00:14 LGRLDR   NA      REC     
#>  7 3DD.007784… 2019-05-02 17:34:35 2019-05-03 11:48:01 GRA      08      INT     
#>  8 3DD.007784… 2019-05-02 17:34:35 2019-05-03 11:48:02 GRA      08      INT     
#>  9 3DD.007784… 2019-05-02 17:34:35 2019-05-03 11:50:16 GRA      06      INT     
#> 10 3DD.007784… 2019-05-02 17:34:35 2019-05-03 11:50:17 GRA      06      INT     
#> # ℹ 67,193 more rows
#> # ℹ 27 more variables: stage <chr>, shuc <dbl>, srkm <chr>, mark_date <date>,
#> #   file_id <chr>, mark_site <chr>, rel_site <chr>, rel_date <date>,
#> #   t_rear_type <chr>, t_species <dbl>, t_run <dbl>, length <dbl>,
#> #   esu_type <chr>, rhuc <dbl>, rrkm <chr>, flags <chr>, trans_status <chr>,
#> #   trans_proj <chr>, trans_year <dbl>, r_flag <chr>, r_rel_site <chr>,
#> #   r_rel_time <dttm>, r_species <dbl>, r_run <dbl>, r_reartype <chr>, …